moved chatmember class to its own file chatmember.domain.ts
