+++ transcendence-app/src/chat/chat.domain.ts
@@ -1,7 +1,6 @@
 /* eslint-disable @typescript-eslint/no-inferrable-types */
 import { Exclude } from 'class-transformer';
 import { ChatEntity } from './infrastructure/db/chat.entity';
-import { ChatMemberEntity } from './infrastructure/db/chatmember.entity';
 
 export class Chat {
   id!: string;
@@ -15,18 +14,6 @@ export class Chat {
   ownerId!: string;
 
   constructor(entity: ChatEntity) {
-    Object.assign(this, entity);
-  }
-}
-
-export class ChatMember {
-  chatId!: string;
-  userId!: string;
-  joinedAt!: Date;
-  admin: boolean = false;
-  muted: boolean = false;
-  banned: boolean = false;
-  constructor(entity: ChatMemberEntity) {
     Object.assign(this, entity);
   }
 }
